Primarily performance improvments to speed up goals.py.
Cleaning up loaded card values on init instead of in each getter.

Also, using the card index value as the __hash__ value so it doesn't
have to be computed off the singular string every time. This seems to
have made an improvement in performance, based on profiler.

Fixed a bug related to achievements and curses.

Modified logging to get it working from the progress meter. Not sure
why the module-level logger wasn't doing the trick.