Fixed some bugs and did some of the TODOs.
Only set the collection member variable when the db is non-null.

Added missing docstring for scrape_and_store_rawgames().

Corrected logic to copy rawgames when they *aren't* present in S3
instead of when they are.

Only insert rawgames when the count in the tarfile is different than
the count in the database. There is still some processing time to read
the compressed tarfile, but less than reading and inserting.

Return the number of games inserted so that the caller can decide
whether there is anything to do next to the date.