Resolves issues with the `/openings` URL.
Obtain the card costs from the appropriate property in the
`dominioncards` module. This was probably broken by 3384b90e.

On the openings page, handle the change to the openings IDs due to the
change to `__repr__()`.

When generating the card names in the dropdown list on `/openings`,
use the `singular` version instead of `str()`, which implicitly calls
`__repr__()`  (e.g., `Card(Chapel)`).

Enables lookup of cards by their `__repr__()` string.

Eliminate a pylint warning about reusing the `c` variable name in
`dominioncards`.

Closes #31.