Added timeout to watch_and_log().
We were encountering problems where a worker would lose connection to
redis, it would complain that it couldn't ack a task, and then Kombu
(used by Celery) would restore it. But the original task seemed to be
left in Pending status forever. To try and handle this situation, the
watch_and_log() function now tries to make sure task statuses are
updating. If they don't, it will resurn after the timeout period. This
is relatively little tested, as no connections were lost in subsequent
runs.

Also tweaked logging a bit.