Accept date range on command line. Some other changes, see commit message.
Interim commit that halts after scraping and loading, before
analyzing. The TODO section around line 104 will need to be removed.

Also reverted the processing to count all the goals before calculating
game stats. This makes it a little easier to get one portion of the
analysis pipeline correct before moving on to the next.