Updates to docs and configs to enable loading data in Vagrant.
The README.rst file was significantly updated to better describe the
steps necessary to create a working development environment.

The councilroom.ini file was moved to /etc/councilroom.ini in the
development guest.

The councilroom-run.yml playbook was modified to try and leave the
councilroom application up and running upon completion.